Renamed hudson.plugins.perforce.security to hudson.plugins.perforce.credentials
