+++ src/main/java/hudson/plugins/perforce/PerforceSCM.java
@@ -27,9 +27,9 @@ import hudson.model.listeners.ItemListener;
 import hudson.plugins.perforce.config.CleanTypeConfig;
 import hudson.plugins.perforce.config.MaskViewConfig;
 import hudson.plugins.perforce.config.WorkspaceCleanupConfig;
-import hudson.plugins.perforce.security.P4CredentialsProvider;
-import hudson.plugins.perforce.security.P4CredentialsProviderDescriptor;
-import hudson.plugins.perforce.security.P4LocalPassword;
+import hudson.plugins.perforce.credentials.P4CredentialsProvider;
+import hudson.plugins.perforce.credentials.P4CredentialsProviderDescriptor;
+import hudson.plugins.perforce.credentials.P4LocalPassword;
 import hudson.plugins.perforce.utils.MacroStringHelper;
 import hudson.plugins.perforce.utils.ParameterSubstitutionException;
 import hudson.remoting.VirtualChannel;
