+++ src/main/java/hudson/plugins/perforce/credentials/P4GlobalPassword.java
@@ -22,7 +22,7 @@
  * THE SOFTWARE.
  */
 
-package hudson.plugins.perforce.security;
+package hudson.plugins.perforce.credentials;
 
 import hudson.Extension;
 import hudson.plugins.perforce.Messages;
