+++ src/main/resources/hudson/plugins/perforce/credentials/P4GlobalPassword/config.jelly
