+++ src/main/resources/hudson/plugins/perforce/credentials/P4LocalPassword/config.jelly
